Fix the CI summary logging of regression tests.
Commit 60c412573efe5b8c51b81925334f758d93d52a5a recently tweaked this
and introduced some errors.
The `&> checkModels.log` syntax overwrites the checkModels.log file
every time, with the combination of stderr and stdout, rather than
appending to it.  Thus some of the intended output of the 
observables testing would be captured, and then only the last iteration
of it would be reported during the "Debugging info" block.

Also, the use of back ticks in the bash echo "<summary>...`checkModels.py`.."
line actually executed the checkModels.py script rather than just
printing its name! But rather than fix this, it is now removed, since
it is no longer used.

The current implementation, hopefully, hides the errors in a summary/details
block, but reports them closer to where they occur.

In general, we should not be hiding or throwing away warnings and error messages!