Remove the DISABLE_JULIA option.
This was only introduced very recently, so probably nobody is using it.
After the changes on this branch, using it only causes an Exception 
anyway, so it's not a useful feature - it's one that just causes a crash!
So I'm removing the feature.
We now only load Julia if we really need it, so disabling it is not 
ever beneficial.